[SPARK-38522][SS] Enrich the method contract of iterator in StateStore to not expect strong consistency on certain condition
### What changes were proposed in this pull request?

This PR proposes to enrich the method contract of iterator in StateStore, that the returned iterator is not guaranteed to reflect all updates being performed "after" it has been initialized.

### Why are the changes needed?

The lack of information on the method contract could mislead callers to expect some guarantees although not described. It would be nice to add to the contract to make clear. It could also help to implementations that they are not required to guarantee such thing.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

N/A

Closes #35816 from HeartSaVioR/SPARK-38522.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>