[SPARK-32268][SQL][FOLLOWUP] Add RewritePredicateSubquery below the InjectRuntimeFilter
### What changes were proposed in this pull request?

Add `RewritePredicateSubquery` below the `InjectRuntimeFilter` in `SparkOptimizer`.

### Why are the changes needed?

It seems if the runtime use in-subquery to do the filter, it won't be converted to semi-join as the design said.

This pr fixes the issue.

### Does this PR introduce _any_ user-facing change?

No, not released

### How was this patch tested?

Improve the test by adding: ensure the semi-join exists if the runtime filter use in-subquery code path.

Closes #35998 from ulysses-you/SPARK-32268-FOllOWUP.

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>