[SPARK-38701][SQL] Inline `IllegalStateException` out from `QueryExecutionErrors`
### What changes were proposed in this pull request?
In the PR, I propose to remove `IllegalStateException` from `QueryExecutionErrors`, and inline such exceptions in the places where they are used.

### Why are the changes needed?
To improve code maintenance. The exceptions of QueryExecutionErrors are supposed to migrate on error classes and be visible to users. `IllegalStateException` shouldn't be propagated to user space.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By existing tests.

Closes #36014 from MaxGekk/throw-IllegalStateException.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>