[MINOR][CORE] Change the log level to WARN for the message which is shown in case users attemp to add a JAR twice
### What changes were proposed in this pull request?

This PR changes the log level to WARN for the message which is shown in case users attempt to add a JAR twice by `sc.addJar`.

### Why are the changes needed?

It's for consistency. `sc.addFile` and `sc.addArchive` show the message as warning in such case.

### Does this PR introduce _any_ user-facing change?

Yes. The message will appear even when the log level of the logger is WARN. But I don't think it affects the behavior.

### How was this patch tested?

Manually confirmed like as follows:
```
sc.addJar("file:////tmp/test.jar")
sc.addJar("file:////tmp/test.jar")
22/02/09 00:56:28 WARN SparkContext: The JAR file:////tmp/test.jar at spark://192.168.1.204:44533/jars/test.jar has been added already. Overwriting of added jar is not supported in the current version.
```

Closes #35443 from sarutak/add-jar-warn.

Authored-by: Kousuke Saruta <sarutak@oss.nttdata.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>