[SPARK-36681][CORE][TESTS][FOLLOW-UP] Tests only when Snappy native library is available in low Hadoop versions
### What changes were proposed in this pull request?

This PR is a minor followup to only test when Snappy native library is available in low Hadoop versions.
From Hadoop 3.3.1 with `HADOOP-17125`, the tests should pass but it fails in lower versions of Hadoop when Snappy native library is unavailable (see also https://github.com/apache/spark/pull/35784#issuecomment-1081290978).

### Why are the changes needed?

To make the tests robust.

### Does this PR introduce _any_ user-facing change?

Nope, this is test-only.

### How was this patch tested?

Should monitor CI

Closes #36136 from HyukjinKwon/SPARK-36681.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>