[SPARK-38730][SPARK-38731][SQL][TESTS] Move tests for error classes of grouping
### What changes were proposed in this pull request?
In the PR, I propose to get together all tests for error classes related to grouping and move them to `QueryCompilationErrorsSuite`.

### Why are the changes needed?
To improve code maintenance - all tests for error classes are placed to Query.*ErrorsSuite. Also exception are raised from `QueryCompilationErrors`, so, tests should be in `QueryCompilationErrorsSuite` for consistency.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By the affected test suites:
```
$ build/sbt "test:testOnly *SQLQuerySuite"
$ build/sbt "test:testOnly *DataFrameAggregateSuite"
$ build/sbt "test:testOnly *QueryCompilationErrorsSuite"
```

Closes #36147 from MaxGekk/move-group-error-class-tests.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>