[SPARK-38589][SQL] New SQL function: try_avg
### What changes were proposed in this pull request?

Add a new SQL function: try_avg. It is identical to the function `avg`, except that it returns NULL result instead of throwing an exception on decimal/interval value overflow.
Note it is also different from `avg` when ANSI mode is off on interval overflows
| Function         | avg                                | try_avg      |
|------------------|------------------------------------|-------------|
| year-month interval overflow | Error                       | Return NULL |
| day-time interval overflow | Error | Return NULL |

### Why are the changes needed?

* Users can manage to finish queries without interruptions in ANSI mode.
* Users can get NULLs instead of runtime errors if interval overflow occurs when ANSI mode is off. For example
```
> SELECT avg(col) FROM VALUES (interval '2147483647 months'),(interval '1 months') AS tab(col)
java.lang.ArithmeticException: integer overflow.

> SELECT try_avg(col) FROM VALUES (interval '2147483647 months'),(interval '1 months') AS tab(col)
NULL
```

### Does this PR introduce _any_ user-facing change?

Yes, adding a new SQL function: try_avg. It is identical to the function `avg`, except that it returns NULL result instead of throwing an exception on decimal/interval value overflow.

### How was this patch tested?

UT

Closes #35896 from gengliangwang/tryAvg.

Lead-authored-by: Gengliang Wang <gengliang@apache.org>
Co-authored-by: Gengliang Wang <ltnwgl@gmail.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>