[SPARK-38581][PYTHON][DOCS] List of supported pandas APIs for pandas-on-Spark docs
### What changes were proposed in this pull request?

This PR proposes to add new page named "Supported pandas APIs" for pandas-on-Spark documents.

### Why are the changes needed?

To let users can more easily find out whether a specific pandas API and its parameters are supported or not from the single document page.

### Does this PR introduce _any_ user-facing change?

Yes, the "Supported pandas APIs" page is added to the user guide for pandas API on Spark documents as below:

<img width="1377" alt="Screen Shot 2022-04-18 at 10 05 30 AM" src="https://user-images.githubusercontent.com/44108233/163739359-972a16bf-b585-4aff-b206-8831592662c4.png">

### How was this patch tested?

Manually check the links in the documents & the existing doc build should be passed.

Closes #36083 from itholic/SPARK-38581.

Authored-by: itholic <haejoon.lee@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>