[SPARK-38970][CI] Skip build-and-test workflow on forks when scheduled
### What changes were proposed in this pull request?
The `configure-jobs` and `precondition` jobs in the `build-and-test` workflow should be skipped on fork repositories as all subsequent / dependent jobs are already skipped.

### Why are the changes needed?
The `build and test` workflow is scheduled to run six times a day. The jobs then only run for `type == 'scheduled'` inside apache/spark repository, not for forks. However, the `configure-jobs` and `precondition` jobs always run, even though all subsequent jobs are skipped on forks.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
No

Closes #36285 from EnricoMi/branch-dont-check-changes-on-forks.

Authored-by: Enrico Minack <github@enrico.minack.dev>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>