[SPARK-34863][SQL][FOLLOW-UP] Handle IsAllNull in OffHeapColumnVector
### What changes were proposed in this pull request?

This PR fixes an issue of reading null columns with the vectorised Parquet reader when the entire column is null or does not exist. This is especially noticeable when performing a merge or schema evolution in Parquet.

The issue is only exposed with the `OffHeapColumnVector` which does not handle `isAllNull` flag - `OnHeapColumnVector` already handles `isAllNull` so everything works fine there.

### Why are the changes needed?

The change is needed to correctly read null columns using the vectorised reader in the off-heap mode.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?

I updated the existing unit tests to ensure we cover off-heap mode. I confirmed that the tests pass with the fix and fail without.

Closes #36366 from sadikovi/fix-off-heap-cv.

Authored-by: Ivan Sadikov <ivan.sadikov@databricks.com>
Signed-off-by: Chao Sun <sunchao@apple.com>