[SPARK-39046][SQL] Return an empty context string if TreeNode.origin is wrongly set
### What changes were proposed in this pull request?

For the query context `TreeNode.origin.context`, this PR proposal to return an empty context string if
* the query text/ the start index/ the stop index is missing
* the start index is less than 0
* the stop index is larger than the length of query text
* the start index is larger than the stop index

### Why are the changes needed?

There are downstream projects that depend on Spark. There is no guarantee for the correctness of TreeNode.origin. Developers may create a plan/expression with a Origin containing wrong startIndex/stopIndex/sqlText.
Thus, to avoid errors in calling `String.substring` or showing misleading debug information, I suggest returning an empty context string if TreeNode.origin is wrongly set. The query context is just for better error messages and we should handle it cautiously.

### Does this PR introduce _any_ user-facing change?

No, the context framework is not released yet.

### How was this patch tested?

UT

Closes #36379 from gengliangwang/safeContext.

Authored-by: Gengliang Wang <gengliang@apache.org>
Signed-off-by: Gengliang Wang <gengliang@apache.org>