[SPARK-38870][SQL][PYSPARK] Make SparkSession.builder return new builder
## What changes were proposed in this pull request?

In Scala, SparkSession.builder returns a new builder each time,
but in Python, it returns the same builder every time, with the
same static options.

Because of this, whenever SparkSession.builder.getOrCreate() is called,
the options set in the very first builder are reapplied.

To fix this, I introduce a classproperty decorator, that is similar
to the property decorator, except that it works with class methods.
Now, whenever `SparkSession.builder` is called, a new builder is
returned as expected.

I also made the parameter `Builder._options` non-static, and I removed
`Builder._sc` which didn't seem used anymore.

`Builder._sc` was introduced in this commit:
e33092d914b6494eadb50a52f7a45509ca36ed5b

And it's only use seems to have been removed by this commit:
3cef320dd99184236660b2863424c36d237223c4

## How was this patch tested?

I updated the Doctests for SparkSession.Builder.getOrCreate and added
Doctests on the classproperty decorator as well.

### What changes were proposed in this pull request?

### Why are the changes needed?

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

Closes #36161 from FurcyPin/FurcyPin/SPARK-38870.

Authored-by: Furcy Pin <furcy.pin@younited-credit.fr>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>