[SPARK-39149][SQL] SHOW DATABASES command should not quote database names under legacy mode
### What changes were proposed in this pull request?

This is a bug of the command legacy mode as it does not fully restore to the legacy behavior. The legacy v1 SHOW DATABASES command does not quote the database names. This PR fixes it.

### Why are the changes needed?

bug fix

### Does this PR introduce _any_ user-facing change?

no change by default, unless people turn on legacy mode, in which case SHOW DATABASES common won't quote the database names.

### How was this patch tested?

new tests

Closes #36508 from cloud-fan/regression.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>