[SPARK-36718][SQL][FOLLOWUP] Fix the `isExtractOnly` check in CollapseProject
### What changes were proposed in this pull request?

This PR fixes a perf regression in Spark 3.3 caused by https://github.com/apache/spark/pull/33958

In `CollapseProject`, we want to treat `CreateStruct` and its friends as cheap expressions if they are only referenced by `ExtractValue`, but the check is too conservative, which causes a perf regression. This PR fixes this check. Now "extract-only" means: the attribute only appears as a child of `ExtractValue`, but the consumer expression can be in any shape.

### Why are the changes needed?

Fixes perf regression

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

new tests

Closes #36510 from cloud-fan/bug.

Lead-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Co-authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>