[SPARK-39086][SQL] Support UDT in Spark Parquet vectorized reader
### What changes were proposed in this pull request?

This PR adds support for `UserDefinedType` when reading Parquet files with the vectorized reader. UDTs mostly consist of data types that vectorized reader could support so it would be good to enable it for UDTs.

### Why are the changes needed?

Enables support for user-defined types in Spark similar to non-vectorized Parquet reader.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

I updated the existing tests to reflect UDT support and added a new test case in ParquetQuerySuite to check UDTs for primitive, struct, and array types.

Closes #36427 from sadikovi/parquet-udt-support.

Authored-by: Ivan Sadikov <ivan.sadikov@databricks.com>
Signed-off-by: Chao Sun <sunchao@apple.com>