[SPARK-28516][SQL] Implement `to_char` and `try_to_char` functions to format Decimal values as strings
### What changes were proposed in this pull request?

This PR implements the `to_char` and `try_to_char` SQL function expressions according to new semantics described below. The former is equivalent to the latter except that it throws an exception instead of returning NULL for cases where the input decimal value does not match the format string.

-----------

# `try_to_char` function

Returns Decimal-typed `expr` cast to STRING using formatting `fmt`, or `NULL` if `expr` is not a valid match for the given format.

## Syntax

```
try_to_char(expr, fmt)
fmt
  { ' [ S ] [ L | $ ]
      [ 0 | 9 | G | , ] [...]
      [ . | D ]
      [ 0 | 9 ] [...]
      [ L | $ ] [ PR | MI | S ] ' }
```

## Arguments

- `expr`: A DECIMAL expression representing a number. `expr` may include leading or trailing spaces.
- `fmt`: An STRING literal, specifying the expected format of `expr`.

## Returns

A STRING representing the result of the formatting operation.

`fmt` can contain the following elements (case insensitive):

```
'0' or '9':

Specifies an expected digit between 0 and 9. A sequence of 0 or 9 in the format
string matches a sequence of digits in the input value, generating a result string
of the same length as the corresponding sequence in the format string. The result
string is left-padded with zeros if the 0/9 sequence comprises more digits than the
matching part of the decimal value, starts with 0, and is before the decimal point.
Otherwise, it is padded with spaces.

'.' or 'D':

Specifies the position of the decimal point (optional, only allowed once).

',' or 'G':

Specifies the position of the grouping (thousands) separator (,). There must be a
0 or 9 to the left and right of each grouping separator.

'$':

Specifies the location of the $ currency sign. This character may only be specified once.

'S' or 'MI':

Specifies the position of a '-' or '+' sign (optional, only allowed once at the
beginning or end of the format string). Note that 'S' prints '+' for positive values
but 'MI' prints a space.

'PR':

Only allowed at the end of the format string; specifies that the result string will
be wrapped by angle brackets if the input value is negative.  ('<1>').
```

If `expr` contains any characters other than `0` through `9` and those permitted in `fmt` a `NULL` is returned.

## Examples

```sql
-- The format expects:
--  * an optional sign at the beginning,
--  * followed by a dollar sign,
--  * followed by a number between 3 and 6 digits long,
--  * thousands separators,
--  * up to two dights beyond the decimal point.
Examples:
> SELECT try_to_char(Decimal(454), '999');
"454"
> SELECT try_to_char(Decimal(454.00), '000.00');
"454.00"
> SELECT try_to_char(Decimal(12454), '99,999');
"12,454"
> SELECT try_to_char(Decimal(78.12), '$99.99');
"$78.12"
> SELECT try_to_char(Decimal(-12454.8), '99,999.9S');
"12,454.8-"
> SELECT try_to_char(Decimal(12454.8), '99,999.9S');
NULL
```

### Why are the changes needed?

The new semantics bring Spark into consistency with other engines and grant the user flexibility about how to handle cases where inputs do not match the format string.

### Does this PR introduce _any_ user-facing change?

Yes.

* The minus sign `-` is no longer supported in the format string (`S` replaces it).
* `MI` and `PR` are new options in the format string.
* `to_number` and `try_to_number` are separate functions with different error behavior.

### How was this patch tested?

* New positive and negative unit tests cover both `to_char` and `try_to_char` functions.
* Query tests update as needed according to the behavior changes.

Closes #36365 from dtenedor/to-char.

Authored-by: Daniel Tenedorio <daniel.tenedorio@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>