[SPARK-39096][SQL][FOLLOW-UP] Fix "MERGE INTO TABLE" test to pass with ANSI mode on
### What changes were proposed in this pull request?

This PR is a minor followup of https://github.com/apache/spark/pull/36445 that fixes the tests to pass with ANSI mode is on. Currently, it fails when ANSI mode is on, (https://github.com/apache/spark/runs/6451765082).

### Why are the changes needed?

To make the tests pass with ANSI on.

### Does this PR introduce _any_ user-facing change?

No, test-only.

### How was this patch tested?

Manually ran the tests.

Closes #36570 from HyukjinKwon/SPARK-39096-followup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>