[SPARK-39225][CORE] Support `spark.history.fs.update.batchSize`
### What changes were proposed in this pull request?
- Add a spark property, `spark.history.fs.update.batchSize`(default 1000), which specifies the batch size for updating new eventlog files.
- For each eventlog scan/parse process, only a max of `spark.history.fs.update.batchSize` jobs are scanned/parsed. This allows the scan process ends within a reasonable time, so new eventlog files(appeared after the scan process starts) can be scanned sooner.

### Why are the changes needed?

Current Spark History Server suffers when there are a large number of eventlog files under eventLog.dir: when a SHS starts, the initial scan may take a long time, and new eventlog files would not be scanned/parsed until the initial scan completes.

For example, if the initial scan takes 1-2 days(this is not uncommon in large environments), the newly finished spark jobs would not show up in SHS since their eventlog files are not scanned/parsed until the initial scan process finishes. This would result in a 1-2 days SHS malfunctioning since the newly finished spark jobs are most likely to be queried by users.

This PR adds a limit for each scan process to `spark.history.fs.update.batchSize`, so that each scan can finish within a reasonable time(default 1000 jobs can usually be scanned in ~10-15 mins but this can vary depending on the sizes of the eventlog files). This will prevent the long initial scan process from blocking the newly appeared eventlog files be scanned.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Tested locally:

- Tested with unit tests in the core module and all passed.
- Placed 10 eventlog files under `eventLog.dir`, and set `spark.history.fs.update.batchSize` to 2. When the SHS started, observed that the scan process parsed 2 files at each time(starting from the files with the latest modification time).
- Added a new eventlog file after the first scan started, and observed that the 2nd scan process picked up and parsed the new eventlog file before parsing other remaining (older) eventlog files.

Closes #36597 from hai-tao-1/updateBatchSize.

Authored-by: Hai Tao <hai_tao@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>