[SPARK-39204][CORE] Change `Utils.createTempDir` and `Utils.createDirectory` call the same logic method in `JavaUtils`
### What changes were proposed in this pull request?
SPARK-39102 introduces two methods `createTempDir` and `createDirectory` into `JavaUtils`, which have the same logic as the method with the same name in `Utils`, in order to continuously ensure they have the same logic,  this pr is change `Utils.createTempDir` and `Utils.createDirectory` call the same logic method in `JavaUtils`.

### Why are the changes needed?
Keep only one `createTempDir()` and `createDirectory()` logic in Spark.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?

- Pass GitHub Action.

Closes #36611 from LuciferYang/SPARK-39204-2.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: Sean Owen <srowen@gmail.com>